[web] Migrate to new id schema
Summary:
https://linear.app/comm/issue/ENG-3795/migrate-the-sqlite-database
https://linear.app/comm/issue/ENG-3796/migrate-redux-state

ID schema migrations for the web. Also changes the `website-responder.js` to always respond with converted ids. This is because when the new web app version releases it expects that the redux store will contain converted ids (with the exception of persisted values, which are migrated).

Depends on D8382

Test Plan:
- Run the old version of the web app, create some drafts, close the web app
- Run the new version of the web app, check if drafts still exists
- Explore the app, check if there are any problems
- Compare redux and sqlite from before and after, and check if all ids were converted

Reviewers: kamil, tomek

Reviewed By: kamil

Subscribers: atul, ashoat

Differential Revision: https://phab.comm.dev/D8383