[lib][native] Remove `displayFailureAlert` prop from `*EditUserAvatarProvider`
Summary:
As of D8341 we no longer use or need this prop. Surfacing of errors is handled by platform-specific code for consistency.

Now we just need to get rid of `useUploadSelectedMedia` and we can consolidate `NativeEditUserAvatarProvider`, `WebEditUserAvatarProvider`, and `BaseEditUserAvatarProvider` into a single platform-agnostic `EditUserAvatarProvider`.

---

Depends on D8341

Test Plan: NA/flow/eslint

Reviewers: ashoat, ginsu, rohan

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8342