Consolidate `[Web/Native/Base]EditUserAvatarProvider` into `EditUserAvatarProvider`
Summary:
As of this diff all of the platform-specific functionality has been moved out of `BaseEditUserAvatarProvider`, which allows us to consolidate the three previous providers into one.

This concludes most of the refactoring work aimed at decoupling `avatar-hooks` and `*AvatarProvider`s.

---

Depends on D8346

Test Plan: Still able to update avatars as expected on both `native` and `web`.

Reviewers: ashoat, ginsu, rohan

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8347