[keyserver] accountOwnershipResponder
Summary:
new responder that checks viewer's username and provides a signed message
attesting that the user owns that particular username. this message can then be
presented to the identity service to allow registration of a reserved username.

Test Plan: called the responder from web and verified that the signature returned with my local keyserver's ed25519 public content key

Reviewers: atul, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8407