Use dedicated functions to get blobURI
Summary:
Part of [[ https://linear.app/comm/issue/ENG-3966/rename-the-holder-message-field-to-bloburi | ENG-3966 ]].

In this diff, I use the new utilities from the previous diff to correctly read the blob URI in all places where it's accessed. This ensures compatibility with both old and new message formats.

Depends on D8372

Test Plan: This stack is tested altogether, actual test plan in the last diff.

Reviewers: ashoat, atul

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8373