[native] Navigate from the community action sheet to the new community roles screen
Summary:
Now that the role navigator is created, we can add the community roles screen and navigate to it when the 'Roles' option is selected on the action sheet. `CommunityRolesScreen` will be implemented later in the stack, this just handles the navigation aspect to make review easier and simpler overall. This screen will display all of the existing roles and member counts within a community, and an option to create a new role with selected permissions.

I created a new `headerLeft` since I couldn't re-use it from the change-roles flow as that had some role-changing specific logic built into it.

Depends on D8335

[[ https://linear.app/comm/issue/ENG-4169/navigate-from-the-community-action-sheet-to-the-new-community-roles | ENG-4169 ]]

Test Plan:
Confirmed that the navigation works and it takes me to the newly created screen.

{F610698}

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8358