[native] Populate the create role screen
Summary:
This diff adds the necessary content to the create roles screen, i.e. a role name input and a panel of selectable permissions for the given role. The `onSave` functionality will be added in a later
diff, as well as the removal of the `eslint-disable-next-line no-unused-vars`.

Depends on D8380

[[ https://linear.app/comm/issue/ENG-4172/populate-the-create-role-screen  | ENG-4172 ]]

Test Plan: Navigated to the screen and typed in different role names, selected several permissions, tried clearing all permissions at once, and tried deselected some permissions individually.

Reviewers: atul, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8391