[clients] Handle notifs with old id schema
Summary:
[ENG-4183](https://linear.app/comm/issue/ENG-4183)

A notif can be sent after the app updates, but before migration is run or the keyserver is notified of the new codeVersion. In this case we need to handle notifications with older ids. Fortunately we can just detect if the id's contain '|' character and if not, we can append the keyserver id.
We don't need to do any c++ changes on the native side, because data is saved to the db immediately, so it will just get migrated normally.

Test Plan:
Test web, iOS, macOS (windows wasn't tested but all desktop changes are on the web app level, so they are platform agnostic):
- try sending a notif with new id schema, check if they are correctly handled
- try sending a notif with old id schema, check if they are correctly handled

Reviewers: marcin, kamil

Reviewed By: marcin

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8460