[Identity] Fix opaque login
Summary:
The DDB query operation coupled with KeyConditionExpression
will project only the PartitionKey and the conditional attribute that
you filtered by. So the resulting query would only return `username` and
`user_id` for a username query.

Re-querying by partition_key allows for the entirity of the object to be
returned.

Depends on D8400

Test Plan:
Login with password.
Tested by Keyserver login diff later on.

Reviewers: varun, bartek, marcin

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8402