[keyserver] call identity service during account creation and deletion
when we create a new account with a username, we should add it to the reserved usernames list
being maintained by the identity service.

when we delete an existing account with a username, we should remove it from the reserved usernames list being maintained by the identity service.

Depends on