[keyserver/lib/native] Create updates for new role creation
Summary: We want to generate updates for the thread when a new role is created and each member of that thread should recieve the update. The next diff will handle the client-side update of pulling the updated
threadInfo from redux (a solution discussed in my 1:1 with Ashoat).

Depends on D8421

Test Plan: Verified that the `threadInfo` was successfully updated in redux after the updates were generated, and checked the thread store operations in the reducer and confirmed that the thread that needed
to be replaced was the one that just had a new role created for it.

Reviewers: atul, ginsu

Subscribers: