[native] Disable headerRight when the role name is empty
Summary: In an effort to prevent the user from reaching a state that would result in an error (and just a general better UX overall), we should disable the `headerRight` 'Create' button when the
role name is empty. We should allow a user to create a role without any selected permissions because on the `keyserver`, we add a set of guaranteed permissions anyways, so we won't end up with a case of a
role with no permissions.

Error handling for duplicate role names / etc will be done in a later diff.

Part of [[ https://linear.app/comm/issue/ENG-4178/support-loadingerror-states-when-creating-a-role | ENG-4178 ]]

Depends on D8428

Test Plan: Verified that the `headerRight` component is disabled when it should be, and enabled when the role name is non-empty.

Reviewers: atul, ginsu

Subscribers: