[native] fix emoji keyboard theme colors
Summary:
Small follow up diff to address comments in D8310.

In this diff I removed all the instances of the panel color family and the emoji keyboard specific colors and committed to using the modal color family. Please note I tried really hard to reuse existing color variables and to not introduce any new variables for this; however, this did lead me to using some variables that didn't "fit" the proper description even though it was the best fitting color. I will highlight this using inline comments and offer an alternate solution if we think the current variable name is not fitting well.

Test Plan:
Please see the screenshots below to confirm that the color themes for both dark and light mode okay. Also had @ted take a look irl

{F623989}

{F623988}

Reviewers: atul, kamil, ashoat

Reviewed By: kamil

Subscribers: ashoat, tomek, ted

Differential Revision: https://phab.comm.dev/D8433