[native] Display loading indicator while role creation is in progress
Summary: As another part to improve the UX when creating a role, it'd be good to display a loading indicator if it's particularly slow on production when the 'Create' button is clicked. Specific error
handling will happen later in the stack.

Part of [[ https://linear.app/comm/issue/ENG-4178/support-loadingerror-states-when-creating-a-role | ENG-4178 ]]

Depends on D8430

Test Plan: Add a delay in the role creation after dispatch and verified the loading indicator appears

Reviewers: atul, ginsu

Subscribers: