[keyserver] Add unique key on thread_name in the roles table
Summary: We want to ensure that for a given community, role names are unique (so we don't have several instances of one role with a different set of permissions). The key is a composite key between `thread`
and `name` because we don't want to prevent the same role name from occuring across several different communities (i.e. 'Moderator' should be allowed in two different communities, but cannot exist twice in
the same community).

This is the `keyserver` validation of preventing this error, the next step will involve adding some kind of client-side notice when they click 'Create' that it is a duplicate role name.

Part of [[ https://linear.app/comm/issue/ENG-4178/support-loadingerror-states-when-creating-a-role | ENG-4178 ]].

Depends on D8431

Test Plan: Tried to create a role with a duplicate name ('Members') and verified that an error throws.

Reviewers: atul, ginsu

Subscribers: