[blob-service] Create business logic abstraction
Summary:
Part of [[ https://linear.app/comm/issue/ENG-4269/implement-the-database-schema | ENG-4269 ]].

Previously, the blob service logic was handled directly in the HTTP handlers (and doubled in the gRPC).
We're gonna introduce cleanup tasks too, so decided to abstract it out.

Depends on D8451.

Test Plan: Tested later in the stack

Reviewers: michal, varun, jon, patryk

Reviewed By: varun

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8453