[keyserver] output keyserver's signing ed25519 public key
Summary:
this script will be run on ashoat's keyserver to get the signing ed25519 public key.
we'll then run the identity service in prod with this key set in an env var so that we can verify messages signed by ashoat's keyserver.

Test Plan: ran the script locally and saw my public key printed to the console

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8448