[keyserver] Use blank target in invite link
Summary:
This page is displayed only on iOS devices. For Safari, this changes invite links experience a bit - the browser shows an alert

{F630049}

asking if a page should be opened instead of showing a banner.

{F630050}

It is also possible that this diff could solve an issue on other browsers https://linear.app/comm/issue/ENG-4089/invite-links-dont-work-on-prod#comment-2e1dff05.

Also included a small change where `inviteLinkUrl` is used instead of hardcoded URL.

Test Plan: Open a link and check if the app is opened (both on Safari and Chrome).

Reviewers: kamil, inka, bartek

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8465