[lib] Add hasMinStateVersion
Summary:
[ENG-4350](https://linear.app/comm/issue/ENG-4350)

Added so landing the id schema migration stack doesn't break dev env. Wasn't sure if it should be merged with `hasMinCodeVersion` but I can't imagine a scenario where we want to check both code version and state version, and it would complicate the logic a bit so I decided against it. But I'm open to changes.

Test Plan: Tested with the next diff

Reviewers: kamil, tomek

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8466