[native] Show a menu button next to each role except admins
Summary:
To prepare for editing and deleting roles, we want to display a menu button next to each role except for admins. It's safe to check by name here since the Admins role cannot be renamed, but it's
worth noting that a more sound check would be what is described in [[ https://linear.app/comm/issue/ENG-4155/add-a-admin-role-column-for-threads | ENG-4155 ]]. For now though, I don't think it makes sense to
prioritize this and will be easy enough to handle down the line.

Part of [[ https://linear.app/comm/issue/ENG-4367/show-a-menu-button-next-to-each-role-except-members-and-admins | ENG-4367 ]].

Depends on D8502

Test Plan: Confirmed that the menu button displays for Members and custom roles, but not Admins

Reviewers: atul, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8522