[landing] clean up competitor comparison UI
Summary: As I was inputting copy for one of the competitors I came across some modals/elements that were not super clean visaully. Spoke with @ted and he gave me some tips on how to clean things up.

Test Plan: Please see the screenshots below to see that each point in the summary was addressed

Reviewers:

Subscribers: