[web] Update the change role icon on web
Summary:
The icon for the change role action item in the members modal is currently a `plus-circle`. After some discussion, we need to use the new `edit-1` icon part of the `CommIcon` component.

`MenuItem` defaults to using a `SWMansionIcon` if a icon name is provided:
```
  let menuItemIcon = iconComponent;
  if (icon) {
    menuItemIcon = <SWMansionIcon size="100%" icon={icon} />;
  }
```

so instead I used the existing `iconComponent` prop to pass in the `CommIcon` we're using.

Test Plan:
The new icon looks as expected

{F632135}

Reviewers: atul, ginsu

Reviewed By: ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8479