[keyserver] Serve intermediate iOS page on a new domain
Summary:
This diff introduces a new url config named `invite_url.json` where an intermediate page is configured. This page is displayed to users on iOS which don't have Comm App installed. We have to serve this page on a different domain than `comm.app` because clicking a link to `comm.app` domain on a page served from `comm.app` doesn't open the app.

The config file can look e.g. like this:
```
{
  "baseDomain": "http://localhost:3000",
  "basePath": "/join/",
  "baseRoutePath": "/join/",
  "https": false,
  "proxy": "none"
}
```

Test Plan: Open `localhost:3000/invite/abc` in Chrome and check if the user was redirected to `localhost:3000/{inviteURL}/invite/abc`.

Reviewers: kamil, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8490