[lib] Handle new id schema for default avatars
Summary: After id schema conversion, the hash of threadID changes, which means the default emoji avatars change. This diff fixes that by detecting if the id is from ashoat's keyserver and if yes treating it as the old schema (hashing only the part after '|').

Test Plan: Log the values, and check if the prefix is correctly removed. Check if genesis default emoji matches the prod one.

Reviewers: kamil, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8491