[native, lib] Move resetUserStateActionType to lib
Summary:
As mentioned in https://linear.app/comm/issue/ENG-4377/refactor-cookie-field-in-web-redux I need to move resetUserStateActionType to lib, because the cookie filed will be in the KesyerveStore, that will be in BaseAppState, so it makes
sense it be reduced by a reducer from lib. And resetUserStateActionType influences the cookie field in redux-setup.js. Since the action will now be in lib, we can move the logic that reduces CurrentUserInfo when this action is called
to lib as well

Test Plan:
ran `yarn flow all`. Checked that `resetUserStateActionType` is not reduced anywhere else.

Reviewers:
michal, kamil

Subscribers: