[native] hide native-specific code
Summary:
Hiding code that will require adding to web dependencies like `olm.h` or `SecureStore` implementation.

NOTE: Using `#ifdef` is not the cleanest and this will be refactored, native-only code will be extracted, but this requires breaking changes and will take some review cycles, that being said I want to port this on web and make it workable (this is the goal) and right after I will put up diffs with refactor.

Depends on D8549

Test Plan: `yarn build-db-wasm`

Reviewers: michal, tomek, jon

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D8550