[lib] Update genesis constant id to new schema
Summary:
[ENG-4394](https://linear.app/comm/issue/ENG-4394/genesis-constant-needs-to-be-updated-to-new-id-schema)

After the id schema migration, the `genesis.id` constant needs to be different on clients, but remain the same on the keyserver. This diff adds a new env var to check if the code is running in the keyserver and if not, changes the genesis id.

Test Plan:
`console.log` on web, keyserver dev and keyserver prod, and check if they are correct

Reviewers: kamil, tomek

Subscribers: