[keyserver] Get rid of code for updating client ThreadStore when upgrade to codeVersion 70 occurs
Summary:
See the first diff in the stack. We're no longer expecting to see such an old `codeVersion`, so I'm going to delete this code.

This is admittedly slightly more risky than some of the other diffs in the stack, because this doesn't just break the USE of such an old `codeVersion`, but rather breaks the upgrade FROM such an old `codeVersion`.

However, I think it's safe because we don't currently have any logged-in sessions with such an old `codeVersion`, and we prevent log in / registration with such an old `codeVersion` in the parent diff. So we don't expect such an upgrade to occur.

Depends on D8503

Test Plan: I tested this whole stack together by playing around with the web app and making sure thread mutation endpoints seemed to work correctly

Reviewers: michal, kamil

Reviewed By: kamil

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8504